requirements.txt
fastapi
uvicorn
openai
PyPDF2
python-dotenv
